Add new component ButtonAdd
